Make stats int

+++ include/stats.hpp
@@ -4,7 +4,7 @@
 #include <string>
 
 void statsSetWorld(const std::string &worldName);
-float statsGetEntry(const std::string &entryKey);
-void statsSetEntry(const std::string &entryKey, float value);
+int statsGetEntry(const std::string &entryKey);
+void statsSetEntry(const std::string &entryKey, int value);
 void statsSave(void);
 void statsLoad(void);
